backend/requirements.txt
fastapi==0.135.1
mysql-connector-python==9.6.0
PyMySQL==1.1.2
python-decouple==3.8
python-dotenv==1.2.2
uvicorn==0.41.0
cryptography
